feat: extract inline components to separate files
Extracted inline components (AnimatedDetails, LoopingText, SocialIcon, ProjectCard) into separate files in src/components for better modularity.
Created src/types/index.ts to centralize shared types (Project, Commit).
Updated src/components/index.ts to export new components.
Refactored src/pages/timeline.tsx, src/components/footer.tsx, src/components/projects.tsx to use extracted components.
