+++ src/components/index.ts
@@ -1,7 +1,11 @@
+export { default as AnimatedDetails } from "./animated-details"
 export { default as CodeBlock } from "./code-block"
 export { default as Footer } from "./footer"
 export { default as Header } from "./header"
 export { default as Layout } from "./layout"
+export { default as LoopingText } from "./looping-text"
+export { default as ProjectCard } from "./project-card"
 export { default as Projects } from "./projects"
 export { default as ScrollToTop } from "./scroll-to-top"
 export { default as SEO } from "./seo"
+export { default as SocialIcon } from "./social-icon"
